backend/requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
sqlalchemy
pathlib